Update draft.md
Proposed a clarification:

I don't think where the retry code is downloaded or authored is what makes this a grey area. Rather, it's any layer of software that sits above the generic HTTP implementation (e.g. JS in browsers, Java/iOS on mobiles) and that may retry on behalf of the end-user.